Rename {sphinx -> build}.py
